Quietly skipping an absent configuration file.
In the current situation (no config file specified on the command
line) the right assumption is that if the config file is missing
it's because the user wants to specify all the arguments on the
command line.  If the user had listed a config file directly as one
of the cli arguments, then the right thing to do would be to
complain.